requirements.txt
langchain
google-cloud-aiplatform
google-generativeai
qdrant-client
pypdf
python-dotenv
streamlit